Extract Types
Extract types into their own file to make dependencies
easier since the given type will be needed in many places.
