+++ src/index.ts
@@ -1,4 +1,4 @@
-export type { Given, GivenOptions, Constructor, Destructor } from './given';
+export type { Given, GivenOptions, Constructor, Destructor } from './given-types';
 export * from './constructor';
 export * from './errors';
 export * from './compat';
